+++ wiki/Changelog.md
@@ -13,6 +13,7 @@ Configuration
   * Status line is configurable in Options menu, switchable from context menu.
   * New user-definable function toggle-tabbar (#1201).
   * Setting ClicksPlaceCursor presets all readline mouse modes.
+  * `MINTTY_DEBUG=C mintty ...` will list loaded config files (#1181).
 
 ### 3.6.3 (18 Dec 2022) ###